requirements.txt
numpy==1.17.2
torch==1.2.0
numpy==1.17.2
opencv-python==4.1.1.26
python-mnist==0.6
torch==1.2.0
